Factor out configure_index
